==> Tracert_Experiments/1.1_allfeatures_train.ipynb <==
# Create a folder named 'pickles' (you can change the folder name as needed)
import os 
folder_name = 'train_test_pickles'
os.makedirs(folder_name, exist_ok=True)

import pandas as pd
import numpy  as np
import json
import matplotlib.pyplot as plt
import seaborn as sns
import sklearn as sk
import warnings
import platform
import sys
import itertools
import random
warnings.filterwarnings("ignore",category=DeprecationWarning)
warnings.simplefilter("ignore")

latency_df = pd.read_pickle('latency_withprobe.pickle')

#remove the null values from the experiment_df 
nan_rows = latency_df[latency_df['last_rtt'].isnull()]
latency_df = latency_df.dropna(subset=['last_rtt'])

model_df = latency_df.copy()

for iteration in range(30):
    test_indices = []
    train_indices = []

        
    array1 = model_df['prb_id'].unique()
    array2 = model_df['dst_id'].unique()

    # Creating all possible pairs
    pairs = list(itertools.product(array1, array2))

    # Randomly selecting 10 pairs
    selected_pairs = random.sample(pairs, 10)

    # Removing selected pairs from the original list
    for pair in selected_pairs:
        pairs.remove(pair)

    # Creating separate lists
    selected_list = selected_pairs
    remaining_list = pairs

    train_dfs = []
    for i,k in remaining_list:
        temp_df = model_df.loc[(model_df['prb_id'] == i) & (model_df['dst_id'] == k)]
                
        # Append the piece to the selected data
        train_dfs.append(temp_df)
                
        train_df = pd.concat(train_dfs)
            
    test_dfs = []
    for i,k in selected_list:
        temp_df = model_df.loc[(model_df['prb_id'] == i) & (model_df['dst_id'] == k)]
                
        # Append the piece to the selected data
        test_dfs.append(temp_df)

        test_df = pd.concat(train_dfs)
        
    
    trainrtt_mean = train_df['last_rtt'].mean()
    trainrtt_std = train_df['last_rtt'].std()
    
    traindist_mean = train_df['distance'].mean()
    traindist_std = train_df['distance'].std()
    
    # normalizing rtt values for train_df
    train_df['normalizzed_rtt'] = (train_df['last_rtt'] - trainrtt_mean) / trainrtt_std
    train_df['normalizzed_distance'] = (train_df['distance'] - traindist_mean) / traindist_std
    
    # normalizing rtt values for test_df
    test_df['normalizzed_rtt'] = (test_df['last_rtt'] - trainrtt_mean) / trainrtt_std
    test_df['normalizzed_distance'] = (test_df['distance'] - traindist_mean) / traindist_std

    # Pickle the train and test dataframes in the 'pickles' folder
    train_df.to_pickle(os.path.join('train_test_pickles', f'train_df_{iteration}.pickle'))
    test_df.to_pickle(os.path.join('train_test_pickles', f'test_df_{iteration}.pickle'))

    print(f'Iteration {iteration + 1} completed.')


# Create a folder named 'model_results' (you can change the folder name as needed)
results_folder = 'model_results'
os.makedirs(results_folder, exist_ok=True)
