==> Tracert_Experiments/6_probe_status.ipynb <==
#Approach One - log into ripe-atlas and retrieve status at timestamp in experiment data 
import requests
import pandas as pd

experiment_df = pd.read_pickle('traceroute.pickle')

def check_probe_connection(row):
    probe_ips = row['from']
    probe_ids = row['prb_id']
    timestamp = row['timestamp']
    
    results = []
    headers = {
        'Content-Type': 'application/json',
        'Accept': 'application/json',
        'X-Api-Key': "e27c24f4-f226-44f4-9a36-c5b05bd33f55"
    }
    
    
    url = f"https://stat.ripe.net/data/atlas-probes/data.json?probe_id={probe_ids}&timestamp={timestamp}&resource={probe_ips}"
            
    try:
        response = requests.get(url,headers=headers)
        response.raise_for_status()
        data = response.json()
                
        if 'error' in data:
            results.append({'Probe ID': probe_ids, 'Status': 'Error', 'Message': data['error']})
        else:
            if data['data']['probes']:
                results.append({'Probe ID': probe_ids, 'Status': 'Connected', 'Message': ''})
            else:
                results.append({'Probe ID': probe_ids, 'Status': 'Disconnected', 'Message': ''})
            
    except requests.exceptions.RequestException as e:
        results.append({'Probe ID': probe_ids, 'Status': 'Error', 'Message': str(e)})
        
    #df = pd.DataFrame(results)
    return results


indices = experiment_df.index.unique()
len(indices)

Status_list = []

for idx in indices[0:20]:
    group = experiment_df[experiment_df.index == idx]
    group_row = group.iloc[0]
    Status = check_probe_connection(group_row)
    Status_list.append(Status)

result_df = pd.DataFrame(Status_list)

# Use apply() method with lambda function to split values into three columns
result_df[['Probe ID', 'Status', 'Message']] = result_df[0].apply(lambda x: pd.Series([x['Probe ID'], x['Status'], x['Message']]))

# Drop the original column
result_df.drop(0, axis=1, inplace=True)
result_df

#Approach two - download the connection logs from ripe-atlas and use them to deduce the connection status of probes 
#when experiment was carried out 

import json

#function for combining json files 
def process_data(json_file_name, probe_id):
    with open(json_file_name, 'r') as file:
        data = json.load(file)
        
    results = {probe_id: []}

    for item in data:
        from_value = item['from']
        to_value = item['to']
        results[probe_id].append({'from': from_value, 'to': to_value})

    return results


probe_id_list = [33627, 51265, 55787, 61357, 15118]

# Dictionary to store combined json files with probe status 
combined_dict = []

# Iterate over servers
for server in probe_id_list:
    server_name = str(server)
    
    # Read JSON log file
    log_file = server_name + '.json'
    
    results = process_data(log_file, server)
        
    # Append to combined dictionary
    combined_dict.append(results)

# Convert the list of dictionaries into a more accessible format
uptime_dict = {}
for d in combined_dict:
    for server_id, uptime_ranges in d.items():
        uptime_dict[server_id] = uptime_ranges
#uptime_dict

# Function to check if the timestamp is within the server's uptime ranges
def is_probe_up(probe_id, timestamp):
    if probe_id in uptime_dict:
        for uptime_range in uptime_dict[probe_id]:
            if uptime_range['to'] is None:
                if uptime_range['from'] <= timestamp:
                    return "connected"
            else:
                if uptime_range['from'] <= timestamp <= uptime_range['to']:
                    return "connected"
    return "disconnected"

# Iterate through the DataFrame and check if the timestamp occurred during the server's uptime
experiment_df['probe_status'] = experiment_df.apply(lambda row: is_probe_up(row['prb_id'], row['timestamp']), axis=1)


outlier = experiment_df[(experiment_df['prb_id'] == 55787) & (experiment_df['new_time'] > '2023-06-01')]
outlier

status = experiment_df[(experiment_df['prb_id'] == 15118) & (experiment_df['new_time'] <= '2023-08-01') & (experiment_df['probe_status'] ==  'disconnected')]
status.tail(10)

# Probe Resource

## Parameters
- **address_v4 (string)**
  - The last IPv4 address that was known to be held by this probe, or null if there is no known address. Note: a probe that connects over IPv6 may fail to report its IPv4 address, meaning that this field can sometimes be null even though the probe may have a working IPv4.

- **address_v6 (string)**
  - The last IPv6 address that was known to be held by this probe, or null if there is no known address.

- **asn_v4 (integer: int64)**
  - The IPv4 ASN if any.

- **asn_v6 (integer: int64)**
  - The IPv6 ASN if any.

- **country_code (string)**
  - An ISO-3166-1 alpha-2 code indicating the country that this probe is located in, as derived from the user supplied longitude and latitude.

- **description (string)**
  - User-defined description of the probe.

- **first_connected (string: date-time)**
  - When the probe connected for the first time (UTC Time and date in ISO-8601/ECMA 262 format).

- **geometry (number: float)**
  - A GeoJSON point object containing the user-supplied location of this probe. The longitude and latitude are contained within the `coordinates` array.

- **id (integer: int64)**
  - The ID of the probe.

- **is_anchor (boolean)**
  - Whether or not this probe is a RIPE Atlas Anchor.

- **is_public (boolean)**
  - If a probe is not public, then certain details, including exact IP addresses, are not returned.

- **last_connected (string: date-time)**
  - When the probe connected for the last time (UTC Time and date in ISO-8601/ECMA 262 format).

- **prefix_v4 (string)**
  - The IPv4 prefix if any.

- **prefix_v6 (string)**
  - The IPv6 prefix if any.

- **status (string)**
  - A JSON object containing:
    - **id:** The connection status ID for this probe (integer [0-3]).
    - **name:** The connection status (string [Never Connected, Connected, Disconnected, Abandoned]).
    - **since:** The datetime of the last change in connection status

- **tags(array):**
- **total_uptime(integer: int64):**
  - Accumulated uptime for this probe in seconds  
- **type(string)**
  - The type of the object



#Approach three - Use the ripe-atlas Probe Object 
from ripe.atlas.cousteau import Probe

probe_id_list = [33627, 51265, 55787, 61357, 15118,26072,30381,15632,1004200,14866]

# Create an empty DataFrame to store probe information
columns = ["ProbeID", "ASN", "CountryCode", "IPAddress", "Status", "Latitude", "Longitude",
           "Tags","Public","Status_since","Description","Prefix_v4",
            ]

probe_list = []

# Loop through each probe ID and retrieve information
for probe_id in probe_id_list:
    probe = Probe(id=probe_id, is_oneoff=False)
    
    
# Extract probe information
    row_data = {
        "ProbeID": probe.id,
        "ASN": probe.asn_v4,
        "CountryCode": probe.country_code,
        "IPAddress": probe.address_v4,
        "Status": probe.status,
        "Anchor": probe.is_anchor,
        "Latitude": probe.geometry['coordinates'][1],
        "Longitude": probe.geometry['coordinates'][0],
        "Tags": probe.tags,
        "Public": probe.is_public,
        "Since": probe.meta_data.get('status_since'),
        "Uptime": probe.meta_data.get('total_uptime'),
        "FirstConnected": probe.meta_data.get('first_connected'),
        "Prefix_v4": probe.prefix_v4,
        "LastConnected": probe.meta_data.get('last_connected')
    }
    
    # Append the row to the list
    probe_list.append(row_data)
    
# Append the row to the DataFrame
probe_df = pd.DataFrame(probe_list)
    

#convert to ordinary time 
from datetime import datetime
def convert_time(datetime_column):
    new_time_col = []
    for i in datetime_column:
        my_datetime = datetime.fromtimestamp(i)
        new_time_col.append(my_datetime)
    return new_time_col

probe_df['Since'] = convert_time(probe_df['Since'])
probe_df['FirstConnected'] = convert_time(probe_df['FirstConnected'])
probe_df['LastConnected'] = convert_time(probe_df['LastConnected'])

#convert to days the uptime of probes
def convert_uptime(uptime_column):
    new_uptime_col = []
    for i in uptime_column:
        my_uptime = i/3600
        days_uptime = my_uptime/24
        new_uptime_col.append(days_uptime)
    return new_uptime_col

probe_df['Uptime(days)'] = convert_uptime(probe_df['Uptime'])


probe_df

# Save the DataFrame to a pickle file
probe_df.to_pickle('probe_data.pickle')

cols_drop = ['Tags','Uptime(days)']
export_df = probe_df.drop(cols_drop, axis=1)
