==> Tracert_Experiments/10_exp_results.ipynb <==
import os
import pandas as pd

def process_csv(file_path):
    df = pd.read_csv(file_path)
    mean_rmse = df['RMSE'].mean()
    mean_denormalized_rmse = df['Denormalized_RMSE'].mean()
    return mean_rmse, mean_denormalized_rmse


# Get the current script's directory
script_directory = os.getcwd()

# Specify the folder containing your CSVs (assuming it's in the same directory as the script)
folder_path = os.path.join(script_directory, 'model_dist_results')

# Specify the output directory
output_directory = os.path.join(script_directory, 'output')

# Create the output directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

output_csv_path = os.path.join(output_directory, 'dist_summary.csv')

results = []

for filename in os.listdir(folder_path):
    if filename.endswith('.csv'):
        file_path = os.path.join(folder_path, filename)
        mean_rmse, mean_denormalized_rmse = process_csv(file_path)
        results.append({'Filename': filename, 'Mean_RMSE': mean_rmse, 'Mean_Denormalized_RMSE': mean_denormalized_rmse})


results_df = pd.DataFrame(results)
results_df.to_csv(output_csv_path, index=False)

# Get the current script's directory
script_directory = os.getcwd()

# Specify the folder containing your CSVs (assuming it's in the same directory as the script)
folder_path = os.path.join(script_directory, 'model_results')

# Specify the output directory
output_directory = os.path.join(script_directory, 'output')

# Create the output directory if it doesn't exist
os.makedirs(output_directory, exist_ok=True)

output_csv_path = os.path.join(output_directory, 'allfeatures_summary.csv')

results = []

for filename in os.listdir(folder_path):
    if filename.endswith('.csv'):
        file_path = os.path.join(folder_path, filename)
        mean_rmse, mean_denormalized_rmse = process_csv(file_path)
        results.append({'Filename': filename, 'Mean_RMSE': mean_rmse, 'Mean_Denormalized_RMSE': mean_denormalized_rmse})


results_df = pd.DataFrame(results)
results_df.to_csv(output_csv_path, index=False)
